3.4-django-testing/django_testing/requirements.txt
django~=5.1
djangorestframework~=3.15.2
django-filter~=24.3
psycopg2-binary

python-dotenv~=1.0.1

pytest~=8.3.2